Fix: Randomizing board when rng is unlucky
Sometimes, rng is unlucky and it isn't possible
to fit the 5th ship.

Cannot just check and restart function within
getting random coord for each ship, because the
higher level function does not know if there was
an issue or if function was restarted.

Had to use try/catch block at higher level and
make sure to return the restarted function.